feat(#869): rename several classes
