+++ src/main/java/org/eolang/jeo/Disassembler.java
@@ -79,7 +79,7 @@ public class Disassembler {
                     new Disassemble(this.target)
                 )
             )
-        ).apply(new BytecodeRepresentations(this.classes).all());
+        ).apply(new BytecodeClasses(this.classes).all());
         stream.forEach(
             terminated -> {
                 try {
